Move ProDOS utils to separate file.
