+++ AppleWin-VS2022.vcxproj.filters
@@ -106,6 +106,9 @@
     <ClCompile Include="source\Configuration\PageSound.cpp">
       <Filter>Source Files\Configuration</Filter>
     </ClCompile>
+    <ClCompile Include="source\ProDOS_Utils.cpp">
+      <Filter>Source Files\Emulator</Filter>
+    </ClCompile>
     <ClCompile Include="source\ParallelPrinter.cpp">
       <Filter>Source Files\Emulator</Filter>
     </ClCompile>
@@ -424,6 +427,12 @@
       <Filter>Source Files\Configuration</Filter>
     </ClInclude>
     <ClInclude Include="source\ParallelPrinter.h">
+      <Filter>Source Files\Emulator</Filter>
+    </ClInclude>
+    <ClInclude Include="source\ProDOS_Utils.h">
+      <Filter>Source Files\Emulator</Filter>
+    </ClInclude>
+    <ClInclude Include="source\ProDOS_FileSystem.h">
       <Filter>Source Files\Emulator</Filter>
     </ClInclude>
     <ClInclude Include="source\Tfe\Pcap.h">
