+++ AppleWinExpress2019.vcxproj
@@ -99,6 +99,8 @@
     <ClInclude Include="source\NTSC_CharSet.h" />
     <ClInclude Include="source\ParallelPrinter.h" />
     <ClInclude Include="source\Pravets.h" />
+    <ClInclude Include="source\ProDOS_FileSystem.h" />
+    <ClInclude Include="source\ProDOS_Utils.h" />
     <ClInclude Include="source\Registry.h" />
     <ClInclude Include="source\RGBMonitor.h" />
     <ClInclude Include="source\Riff.h" />
@@ -179,6 +181,7 @@
     <ClCompile Include="source\FourPlay.cpp" />
     <ClCompile Include="source\FrameBase.cpp" />
     <ClCompile Include="source\MockingboardCardManager.cpp" />
+    <ClCompile Include="source\ProDOS_Utils.cpp" />
     <ClCompile Include="source\RGBMonitor.cpp" />
     <ClCompile Include="source\SAM.cpp" />
     <ClCompile Include="source\Debugger\Debug.cpp" />
